make forklift_versions a module, not a lookup plugin
it's so much easier to handle it when it's a module
