+++ roles/forklift_versions/tasks/upgrade.yml
@@ -1,10 +1,15 @@
 ---
 - name: get forklift upgrade versions
-  set_fact:
-    forklift_upgrade_versions: "{{ lookup('forklift_versions', 'file='+forklift_versions_file+' scenario='+scenario+' scenario_version='+scenario_version+' scenario_os='+scenario_os+' upgrade=true') }}"
+  forklift_versions:
+    file: "{{ forklift_versions_file }}"
+    scenario: "{{ scenario }}"
+    scenario_version: "{{ scenario_version }}"
+    scenario_os: "{{ scenario_os }}"
+    upgrade: True
+  register: forklift_upgrade_versions
 
 - name: set individual upgrade vars
   set_fact:
-    forklift_upgrade_version_start: "{{ forklift_upgrade_versions[0] }}"
-    forklift_upgrade_version_intermediate: "{{ forklift_upgrade_versions[1] }}"
-    forklift_upgrade_version_final: "{{ forklift_upgrade_versions[2] }}"
+    forklift_upgrade_version_start: "{{ forklift_upgrade_versions.versions[0] }}"
+    forklift_upgrade_version_intermediate: "{{ forklift_upgrade_versions.versions[1] }}"
+    forklift_upgrade_version_final: "{{ forklift_upgrade_versions.versions[2] }}"
