+++ roles/forklift_versions/tasks/versions.yml
@@ -1,8 +1,12 @@
 ---
 - name: get forklift versions
-  set_fact:
-    forklift_versions: "{{ lookup('forklift_versions', 'file='+forklift_versions_file+' scenario='+scenario+' scenario_version='+scenario_version+' scenario_os='+scenario_os) }}"
+  forklift_versions:
+    file: "{{ forklift_versions_file }}"
+    scenario: "{{ scenario }}"
+    scenario_version: "{{ scenario_version }}"
+    scenario_os: "{{ scenario_os }}"
+  register: forklift_versions
 
 - name: set individual component vars
   set_fact: "{{ item.key }}={{ item.value }}"
-  with_items: "{{ forklift_versions | dict2items }}"
+  with_items: "{{ forklift_versions.versions | dict2items }}"
